+++ ContentPatcher/ModEntry.cs
@@ -52,7 +52,8 @@ namespace ContentPatcher
             new Migration_1_16(),
             new Migration_1_17(),
             new Migration_1_18(),
-            new Migration_1_19()
+            new Migration_1_19(),
+            new Migration_1_20()
         };
 
         /// <summary>The special validation logic to apply to assets affected by patches.</summary>